requirements.txt
asgiref==3.6.0
astroid==2.15.4
attrs==23.1.0
# backports.zoneinfo==0.2.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
Deprecated==1.2.13
dill==0.3.6
Django==4.2.1
django-braces==1.15.0
django-cors-headers==3.14.0
django-filter==23.2
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.5.1
drf-yasg==1.21.5
facebook-sdk==3.1.0
Faker==18.7.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==1.4.2
lazy-object-proxy==1.9.0
Markdown==3.4.3
MarkupSafe==2.1.2
mccabe==0.7.0
mysql==0.0.3
mysqlclient==2.1.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.5.0
protobuf==4.22.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests
rsa==4.9
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.9
# urllib3
whitenoise==6.4.0
wrapt==1.15.0
zipp==3.15.0
